# ENH/PERF: optional mask in nullable dtypes