@jorisvandenbossche 2021-06-17
@mzeitlin11 thanks a lot for looking into this!

There is one problem I recently realized while thinking this through (but didn't yet comment about that on the issue), and that is: how to ensure "views" still work?

So for example, assume I create an array with no NAs (the mask will be None), then I create a view of this array (with a slice, or the `view()` method), and then it is expected that modifications in the one array will be reflected in the other array. But when introducing missing values, the mask gets created on the view, and this doesn't propagate to the parent (or vice versa):

```
In [7]: arr = pd.array([1, 2, 3, 4])

In [8]: arr2 = arr[:]

In [9]: arr2[0] = pd.NA

In [10]: arr2
Out[10]: 
<IntegerArray>
[<NA>, 2, 3, 4]
Length: 4, dtype: Int64

In [11]: arr   # <-- not mutated, but should have been
Out[11]: 
<IntegerArray>
[1, 2, 3, 4]
Length: 4, dtype: Int64
```

One possible solution would be to create a small `Mask` class that is stored on the array instead of `np.ndarray | None`, and this object could then actually be shared between views. And it's this object that could "hide" whether the mask is present as an ndarray or not, and if the object gets modified to add an ndarray as mask (eg when setting a NA), then the change can be reflected in the all arrays that share this `Mask` object. 

Such a class could also have some helper methods then like `has_na()`, `maybe_has_na()` (only check if the mask is not None), or `get_ndarray_mask()`, etc.

However, one problem with this (that I only realize now while writing it down), is that if using such a shared Mask object, we would need to keep track of the offset/length on the Array class in case the array is sliced (because the mask would present the full mask of the parent array, and when asking for the mask array or whether there are NAs present, we of course only need the sliced part), which seems like a considerable complication ..

Another option would be to only support "read-only views" (so that if an array is a view of another, it can't be mutated), but that's certainly something that would need to be discussed. 

cc @jbrockmendel @TomAugspurger