@mzeitlin11 2021-06-17
> There is one problem I recently realized while thinking this through (but didn't yet comment about that on the issue), and that is: how to ensure "views" still work?

Thanks for bringing this up, that's a serious complication that never crossed my mind :)

For now I'll try messing around with some workarounds like you mention, though I worry they might complicate the implementation too much to be worth it.