@TomAugspurger 2021-06-19
I think the only way to handle this properly is for the second array to be a view on the original IntegerArray object, rather than it's component ndarrays like we do in `__getitem__` today. That seems like the only way to ensure that the state is actually ensured to be consistent.

I don't know exactly what this would look like in code, but maybe some kind of `.base` property with a type `Optional[IntegerArray]`. Then any references to `._data` or `._mask` would go through `.base` if it's present.